Napredno brisanje Amenity i Comment, dodavanje User, Podeljena pretraga svih Amenity i Comments

+++ src/services/AmenityService.java
@@ -4,6 +4,7 @@ import java.util.Collection;
 
 import javax.annotation.PostConstruct;
 import javax.servlet.ServletContext;
+import javax.ws.rs.Consumes;
 import javax.ws.rs.DELETE;
 import javax.ws.rs.GET;
 import javax.ws.rs.POST;
@@ -14,6 +15,7 @@ import javax.ws.rs.core.Context;
 import javax.ws.rs.core.MediaType;
 
 import DAO.AmenityDAO;
+import DAO.ApartmentDAO;
 import beans.Amenity;
 
 
@@ -33,6 +35,18 @@ public class AmenityService {
 			String contextPath = ctx.getRealPath("");
 			ctx.setAttribute("amenitiesDAO", new AmenityDAO(contextPath));
 		}
+		if(ctx.getAttribute("apartmentDAO") == null) { //TODO da li ovo treba ovde?
+			String contextPath = ctx.getRealPath("");
+			ctx.setAttribute("apartmentDAO", new ApartmentDAO(contextPath));
+		}
+	}
+	
+	@GET
+	@Path("/all")
+	@Produces(MediaType.APPLICATION_JSON)
+	public Collection<Amenity> getAllAmenities(){
+		AmenityDAO dao = (AmenityDAO) ctx.getAttribute("amenitiesDAO");
+		return dao.findAll();
 	}
 	
 	@GET
@@ -40,11 +54,11 @@ public class AmenityService {
 	@Produces(MediaType.APPLICATION_JSON)
 	public Collection<Amenity> getAmenities(){
 		AmenityDAO dao = (AmenityDAO) ctx.getAttribute("amenitiesDAO");
-		return dao.findAll();
+		return dao.findAllUndeleted();
 	}
 	
 	@POST
-	@Path("/")
+	@Path("/write")
 	@Produces(MediaType.APPLICATION_JSON)
 	public void writeAmenities(){
 		AmenityDAO dao = (AmenityDAO) ctx.getAttribute("amenitiesDAO");
@@ -55,7 +69,22 @@ public class AmenityService {
 	@Path("/{id}")
 	@Produces(MediaType.APPLICATION_JSON)
 	public Amenity delete(@PathParam("id") long id) {
+		ApartmentDAO apartmentDAO = (ApartmentDAO) ctx.getAttribute("apartmentDAO");
 		AmenityDAO dao = (AmenityDAO) ctx.getAttribute("amenitiesDAO");
+		
+		if(dao.findAmenity(id)== null) return null;
+		
+		apartmentDAO.deleteAmenity(id);
+		apartmentDAO.write();
 		return dao.delete(id);
+	}
+	
+	@POST
+	@Path("/")
+	@Consumes(MediaType.APPLICATION_JSON)
+	@Produces(MediaType.APPLICATION_JSON)
+	public Amenity newAmenity(Amenity amenity) {
+		AmenityDAO dao = (AmenityDAO) ctx.getAttribute("amenitiesDAO");
+		return dao.save(amenity);
 	}
 }
